Added new interface IContentTypeVersionPolicySupport and implemented it in portal_repository.  It allows products to register versioning policies (currently just a title and id), and to associate those policies with specific portal types.
Added a versioning policy (at_edit_autoversion) which automatically creates new versions on edit for AT types which are configured to support the policy in the configlet.  This is implemented using a simple macro override on AT's edit_macros, and a new entry in the AT edit form controller chain.

svn path=/CMFEditions/branches/alecm-versioning_policies/; revision=11306