portal_repository.listPolicies: log and ignore Broken VersionPolicies.
I want to remove the Archetypes specific ATVersionOnEditPolicy.
It looks like we can do this safely without getting errors.
At some point it is better to delete the object from the repository tool, but it is okay for now.
We can do that in plone.app.upgrade for Plone 6.

For now, edit listPolicies to ignore this policy and log it.
Note that listPolicies is not actually called anywhere as far as I see.

Sample log line (split into three here for readability):

```
INFO [Products.CMFEditions.CopyModifyMergeRepositoryTool:208][waitress-2]
Ignoring broken policy at_edit_autoversion:
<persistent broken Products.CMFEditions.VersionPolicies.ATVersionOnEditPolicy instance b'\x00\x00\x00\x00\x00\x00\x00`'>
```

Update: I actually no longer want to remove the ATVersionOnEditPolicy class,
because it is still needed (see future commit) but checking for Broken policies still seems good.