Allows the "Plone Site" type to be cloned
Products.CMFEditions uses the pickle module to clone objects. See:

https://github.com/plone/Products.CMFEditions/blob/c80bc31af46bff45fee2908878b1f01190fda8d8/Products/CMFEditions/ArchivistTool.py#L210-L229

These changes prevent the error:

TypeError: Can't pickle objects in acquisition wrappers

When trying to serialize an ImplicitAcquisitionWrapper object with the
pickle dump. This error occurred when trying to check in a Plone Site
type with plone.app.iterate.

These changes also allow the Plone Site type to be serialized with the
pickle.